+++ diskover-web/public/js/search.js
@@ -199,9 +199,9 @@ function hideCharts() {
 }
 
 function getJSONFileTree() {
-    console.time('loadtime')
+    console.time('treeloadtime')
     // get json data from Elasticsearch using php data grabber
-    console.log("grabbing json data from Elasticsearch");
+    console.log("grabbing json data from Elasticsearch for file tree");
 
     // config references
     chartConfig = {
@@ -241,7 +241,7 @@ function getJSONFileTree() {
         $("#dirdetails").append(dirdetails);
         // stop spin.js loader
         spinner.stop();
-        console.timeEnd('loadtime');
+        console.timeEnd('treeloadtime');
         // load file tree
         if (getCookie('hidesearchtree') != 1) {
             updateTree(root, root);
@@ -328,6 +328,16 @@ function updateTree(data, parent) {
             return " " + format(d.size);
         })
 
+    //add bar for size percent
+    entered.append("span")
+    .html(function (d) {
+        var barpercent = (d.size / root.size * 100 / 1.5).toFixed(1) + '%';
+        var ret = "<div style=\"position:relative; top:-3px; z-index:-99; height:4px; width:" + barpercent + "; background-image: linear-gradient(to right, #202327, #DB3A13);\"> </div>";
+        var barpercent = (d.count / root.count * 100 / 1.5).toFixed(1) + '%';
+        ret += "<div style=\"position:relative; top:-3px; z-index:-99; height:4px; width:" + barpercent + "; background-image: linear-gradient(to right, #202327, #FF9901);\"> </div>";
+        return ret;
+    });
+
     //update position with transition
     nodeEls.transition().duration(treeduration)
         .style("top", function (d) {
@@ -362,11 +372,12 @@ $(function () {
             .attr('class', 'd3-tip')
             .html(function (d) {
                 var rootval = root.size;
-                var percent = (d.size / rootval * 100).toFixed(1) + '%';
+                var sizepercent = (d.size / rootval * 100).toFixed(1) + '%';
                 var sum = format(d.size);
-                var ret = "<span style='font-size:12px;color:white;'>" + d.name + "</span><br><span style='font-size:12px; color:red;'>" + sum + " (" + percent + ")</span>";
+                var itempercent = (d.count / root.count * 100).toFixed(1) + '%';
+                var ret = "<span style='font-size:12px;color:white;'>" + d.name + "</span><br><span style='font-size:12px; color:#DB3A13;'>" + sum + " (" + sizepercent + ")</span>";
                 if (d.type === 'directory') {
-                    ret += "<br><span style='font-size:11px; color:lightgray;'>Items: " + numberWithCommas(d.count) + "</span>";
+                    ret += "<br><span style='font-size:12px; color:#FF9901;'>Items: " + numberWithCommas(d.count) + " (" + itempercent + ")</span>";
                     ret += "<br><span style='font-size:11px; color:lightgray;'>Files: " + numberWithCommas(d.count_files) + "</span>";
                     ret += "<br><span style='font-size:11px; color:lightgray;'>Folders: " + numberWithCommas(d.count_subdirs) + "</span>";
                 }
@@ -396,9 +407,13 @@ $(function () {
     /* ------- SEARCH RESULTS CHARTS -------*/
 
     if (getCookie('hidesearchcharts') != 1) {
+        console.time('chartloadtime')
+        // get json data from Elasticsearch using php data grabber
+        console.log("grabbing json data from Elasticsearch for chart");
+
         // config references
         var chartConfig = {
-            target: 'searchCharts-container',
+            target: 'charts-container',
             data_url1: 'd3_data_search_dirsizechart.php?path=' + encodeURIComponent(path) + '&usecount=0&usecache=' + usecache,
             data_url2: 'd3_data_search_dirsizechart.php?path=' + encodeURIComponent(path) + '&usecount=1&usecache=' + usecache,
             data_url3: 'd3_data_pie_ext_searchresults.php?path=' + encodeURIComponent(path) + '&usecache=' + usecache,
@@ -415,6 +430,7 @@ $(function () {
             speed: 1.9, // Rounds per second
             trail: 40, // Afterglow percentage
             className: 'spinner', // The CSS class to assign to the spinner
+            position: 'relative'
         };
 
         // loader settings
@@ -422,6 +438,11 @@ $(function () {
 
         // trigger loader
         var spinner = new Spinner(opts).spin(target);
+
+        console.log(chartConfig.data_url1)
+        console.log(chartConfig.data_url2)
+        console.log(chartConfig.data_url3)
+        console.log(chartConfig.data_url4)
 
         // load json data from Elasticsearch
         // use d3 queue to load json files simultaneously
@@ -443,6 +464,7 @@ $(function () {
 
                 // stop spin.js loader
                 spinner.stop();
+                console.timeEnd('chartloadtime');
                 renderCharts(data1, data2, data3, data4);
             });
     }
@@ -663,7 +685,6 @@ $(function () {
 
         // hide top dir chart if no files
         if (data3.length === 0) {
-            $('#mtime-Chart-container').hide();
             $('#topFileTypes-Chart-container').hide();
         } else {
             // by size bar chart
@@ -862,6 +883,12 @@ $(function () {
             //- END BAR/PIE EXT CHART -
             //--------------------
 
+        }
+
+        // hide charts if no files
+        if (data4.length === 0) {
+            $('#mtime-Chart-container').hide();
+        } else {
 
             //----------------
             //- STACKED BAR MTIME CHART -