+++ diskover-web/public/savesearchfilters.php
@@ -17,8 +17,6 @@ https://www.diskoverdata.com/solutions/
 */
 
 require '../vendor/autoload.php';
-
-error_reporting(E_ALL ^ E_NOTICE);
 require "../src/diskover/Diskover.php";
 
 $errors = [];